Images are stored in MongoDB.

+++ backend/seed.js
@@ -4,7 +4,6 @@ import Session from "./models/Session.js";
 import crypto from "crypto";
 import dotenv from 'dotenv';
 import mongoose from "mongoose";
-import imageDownloader from "image-downloader";
 import path from "path";
 import fs from "fs";
 import fetch from "node-fetch";
@@ -24,30 +23,31 @@ const DeleteAll = async () => {
     try {
         fs.readdir(path.resolve() + "/storage/UserCreatedContent/", (err, files) => {
             if (err) {
-                fs.mkdirSync("../storage/UserCreatedContent", {recursive: true});
+                fs.mkdirSync("/storage/UserCreatedContent", { recursive: true });
             }
-            console.log(files);
-            files.forEach((file) => {
-                if (!(protectedFiles.includes(file))) {
-                    try {
-                        fs.unlink(path.resolve() + "/storage/UserCreatedContent/" + file, (err) => {
-                            if (err) {
-                                console.log(`Unable to delete ${path.resolve() + "/storage/UserCreatedContent/" + file}.`);
-                            }
-                            else {
-                                console.log(`Deleted ${path.resolve() + "/storage/UserCreatedContent/" + file}.`);
-                            }
-                        });
+            if (files) {
+                files.forEach((file) => {
+                    if (!(protectedFiles.includes(file))) {
+                        try {
+                            fs.unlink(path.resolve() + "/storage/UserCreatedContent/" + file, (err) => {
+                                if (err) {
+                                    console.log(`Unable to delete ${path.resolve() + "/storage/UserCreatedContent/" + file}.`);
+                                }
+                                else {
+                                    console.log(`Deleted ${path.resolve() + "/storage/UserCreatedContent/" + file}.`);
+                                }
+                            });
+                        }
+                        catch (e) {
+                            console.log(`Unable to delete ${path.resolve() + "/storage/UserCreatedContent/" + file}.`);
+                        }
                     }
-                    catch (e) {
-                        console.log(`Unable to delete ${path.resolve() + "/storage/UserCreatedContent/" + file}.`);
-                    }
-                }
-            });
+                });
+            }
         });
     }
     catch (e) {
-        fs.mkdirSync("/storage/UserCreatedContent");
+        fs.mkdirSync("/storage/UserCreatedContent", { recursive: true });
     }
     await User.deleteMany({});
     await Post.deleteMany({});
@@ -69,17 +69,11 @@ console.log("Setup seed user.");
 console.log("Setting up 20 posts.");
 const SetupPosts = async () => {
     const author = await User.findOne({});
-    async function downloadImage(url, filepath) {
-        try {
-            return imageDownloader.image({
-                url,
-                dest: filepath
-            });
-        }
-        catch (e) {
-            console.log(e);
-            process.exit();
-        }
+    async function downloadImage(url) {
+        const imageResponse = await fetch(url);
+        const imageBlob = await imageResponse.blob();
+        const buffer = await imageBlob.arrayBuffer();
+        return Buffer.from(buffer, "hex");
     }
     async function getQuote() {
         const url = "https://api.muetab.com/quotes/random?language=English";
@@ -90,12 +84,12 @@ const SetupPosts = async () => {
         const quote = quoteJSON.quote;
         return quote;
     }
-    for (let i = 0; i < 20; i++) {
+    for (let i = 0; i < 5; i++) {
         const key = crypto.randomBytes(32).toString("hex");
         const quote = await getQuote();
-        await downloadImage("https://picsum.photos/400/200", path.resolve() + `/storage/UserCreatedContent/${key}.jpg`);
-        const NewPost = new Post({ authorUsername: username, authorID: author._id, text: quote, imageKey: key + ".jpg", date_created: new Date() });
-        NewPost.save();
+        const image = await downloadImage("https://random.imagecdn.app/400/200");
+        const NewPost = new Post({ authorUsername: username, authorID: author._id, text: quote, image: image, date_created: new Date() });
+        await NewPost.save();
     }
 };
 await SetupPosts();
